Introduce 7 static memory allocation policies
This commit introduces 7 static memory allocation policies to
the backend of ArgoDSM, namely cyclic, cyclic_block, skew_mapp,
skew_mapp_block, prime_mapp, prime_mapp_block and first_touch.
The block implementations can address different granularities
other than handling data at the default granularity level (4KB).

* Make use of the global_ptr template class

This commit makes use of the global_ptr template class defined
in data_distribution.hpp and introduces the corresponding
implementation file (data_distribution.cpp), in order to improve
readability by hiding the implementation details of the policies.

* Introduce 2 environment variables

The memory policy, as well as the granularity for the block
implementations, is chosen through the use of the environment
variables ARGO_ALLOCATION_POLICY and ARGO_ALLOCATION_BLOCK_SIZE.
The former variable should be given a number from 0 to 7,
starting from bind_all and continuing with the rest of the
policies in the order presented above.

* Implementation details

Modification of the global_ptr(T* ptr) constructor. The
initial form of the constructor invokes both the homenode
and local_offset functions for the given address. This is
unnecessary to do when only one of two is actually exploited.
Because of that, we introduce an additional input parameter
to the constuctor, which can be used to point to the member
function needed to be invoked.

An additinal parameter is also added to the getHomenode and
getOffset functions, and is introduced because of the internals
of the first_touch memory policy in conjuction with where the
two functions are called inside the handler. Making use of this
parameter essentially helps us keep the original format of the
handler and avoid moving the sem_wait(&ibsem) (line 354) before
the invocation of the getHomenode and getOffset (line 344 & 345).

For the first_touch memory policy, we make use of a directory to
keep track of the owner of every page. The directory memory, as
well as the MPI window that this memory is associated with, are
only constructed when the first_touch memory policy is chosen.