Various changes due to issue #52 and a first-touch limitation
As per issue #52, changes were made to mempool and the collective
allocators, in order for only the home node of the data structure
allocated to apply changes on it (initialization/modification).
As the above didn't solve all issues, further modifications were
applied on the failing tests. More specifically:

- PerfectForwardingCollectiveNew under allocators.cpp
In this test, a global pointer object is attached to the collecti-
vely allocated tuple, in order to identify the home node of this
page and allow only that process to proceed on the modifications.

- selectiveArray, selectiveUnaligned, writeBufferLoad under backend.cpp
In these tests, only node_0 was set to handle the initialization
of the global data structures worked on, thus leading to failure
when utilizing the first-touch memory policy, as the backing store
of node_0 is not sufficient to host all pages. To spread the pages
across all nodes and avoid this issue, the initialization is done
collectively. The ArgoDSM global memory size is also increased
from 16 to 32MB, as writeBufferLoad was still failing even by team
process initializing the data structures.

- ReadUninitializedSinglenode under uninitialized.cpp
This test was also modified to collectively initialize the data
structure worked on due to the aforementioned limitation.

The case where if one node under first-touch initializes all data
leads to failure is left to be handled on future work, by moving
the allocations to the next node of sufficient backing store size.