Merge the single and block page policy implementations
In this commit, we merge the policies operating on a
single page granularity level with their correspondi-
ng block implementations. As the default page block
size is 16, it is left up to the user to reduce it or
increase it, based on his needs and system configura-
tion. For code maintenance, this is better as it eli-
minates code duplication and reduces code complexity.