Fix race in first-touch and update tutorial.md
Under the first-touch policy, a node after failing to claim
ownership of a page, it might be fast enough to fetch the
offset value for this page before the update from the remo-
te node has been reflected back to its local window. For
that reason, we also fill the directory offset indices with
magic values at initialization, in order to spin on these
values later on, and make sure we read the correct offsets.