Revert ReadUninitializedSinglenode back to its original format
Removed team process reads in the ReadUninitializedSinglenode
test and let node_0 do the work as in the original format. The
size when utilizing the first-touch memory policy is simply
narrowed down for the pages to fit in the backing store of no-
de_0. While at it, added @todo comments also in backend.cpp.