Improve first-touch data distribution
This commit improves the first-touch data distribution so that it
passes ownership of a page to a remote node, when its backing sto-
re is not of sufficient size to host more pages. Briefly, the node
first checks if it has sufficient backing store size, and then ri-
ght-wise checks for availability in the rest of the nodes.

For that to be feasible, a global table is introduced, which holds
the backing offsets of the nodes. Additionally, a new column is
introduced to the owners directory, named the `ownership` field.

This commit also:
* Enables back pre-fetching for the first-touch distribution.
* Decentralizes the ownership CAS operation done in first_touch().
* Detaches the directory actions from homenode() and local_offset().
* Performs MPI RMA operations to the public windows in bulk.
* Limits these operations to a specific node, rather than broadcasting.