Fix erroneous write misses
This commit ensures that simultaneous read misses on the same
address do not result in one or more erroneous write misses. This
previously occurred when threads in the signal handler had waited
for the cache mutex and eventually found the page they were
attempting to read to already be fetched. This scenario was
wrongfully interpreted as a write miss.