Add six coherence tests to backendTests
The tests test ArgoDSM coherence by writing to random indices on a set
of pages (first set of tests) and within a page (other set of tests).

Random accesses within the range of several pages:
- randAccessesBarrierArray
- randAccessesBulkySelectiveReleaseAcquireArray
- randAccessesPeriodicSelectiveReleaseAcquireArray

Random accesses within a page:
- randAccessesBarrierPage
- randAccessesBulkySelectiveReleaseAcquirePage
- randAccessesPeriodicSelectiveReleaseAcquirePage

Coherence operations for each test type:
- randAccessesBarrier*
  -- uses an Argo barrier.
- randAccessesBulkySelectiveReleaseAcquire*
  -- uses a bulky selective-release, an MPI barrier, and an acquire.
- randAccessesPeriodicSelectiveReleaseAcquire*
  -- uses periodic selective-release, an MPI barrier, and an acquire.

NOTE: The tests use a constant seed in order to be deterministic.