Disable argo atomics of size other than 4/8 bytes (#88)
This commit disables the use of ArgoDSM atomics for types of size
other than 4 or 8 bytes. UCX only supports 4 and 8 byte atomics and
while a fallback to active messaging is supposed to exist, this
does not always appear to be true.
Also disables the tests using atomics on chars (of size 1 byte).
The relevant code is left commented out for the time being.