Completely clean state for each test group (#83)
This PR changes the behavior of argo::reset to, in addition to resetting
the allocators, also reset Argo's coherence state.
argo::reset is now called ahead of running each test to ensure that each
test executes from a clean state. This helps us expose bugs previously
hidden due to the dirty state of the different directories in the backend.