Align all memory allocations (#90)
This PR ensures that all memory allocations are aligned to at
least the alignment of std::max_align_t. This alignment is
typically 8 or 16 bytes. This is done by allocating in chunks of
alignof(std::max_align_t).

Two new tests to test alignment of allocations were also added.
The first tests that all allocations are aligned as expected, and the
second tests that reusing deallocated memory from the freelist
works as intended with aligned allocations.