Replace std::mutex with QD lock in write buffer (#79)
* Add QD library as submodule under extern/qd.

* Use QD lock in write buffer

Use of QD locking instead of std::mutex for the ArgoDSM write buffer.
Additionally, the public API functionality to add to, erase from and
flush the write buffer delegates the corresponding operations to the
current lock holder.

* Make sure RPATH is set for install targets

With libargobackend-mpi depending on libqd, libqd becomes an
indirect dependency of any executable or library that links with
libargobackend-mpi. libargobackend-mpi must therefore itself be
aware of where to find libqd during execution. This is done by
ensuring that CMake sets the RPATH for each install target so
long as they are not in a system directory.

* Use absolute https path to clone the QD library

Using the absolute path ensures that cloning the QD library is
more flexible if ArgoDSM is not cloned from GitHub.

* Add statistics for awaiting write buffer flush

Measure how much time is spent waiting for the write buffer lock
holder to flush the buffer.  The time can be retrieved by calling
write_buffer::get_flush_wait_time()

* Better explanations for two statistics counters.