Specify which VM model to enable via an = option
Instead of employing `-DARGO_VM_???` options for VM management
(and have complicated code to check that exactly one of them is enabled).

Besides simplifying the code, this will enable testing all options more
easily on GitHub actions.

For the moment, the same default has been kept, but it will need to be
revised when #110 is appropriately fixed.