Eliminate all uses of the magic 4096 constant
This constant is used for the size of a page, but was hard-coded in
various places using various forms (4096, 4096UL, ...).  Also, various
header files used different definitions / names for it (pagesize,
page_size, PAGESIZE).

This commit adds a `PAGE_SIZE` constexpr and uses it in all `src` files.

Addresses the biggest part of #49.  (Does not deal with CACHELINE.)