Pass PAGE_SIZE as a define instead
To avoid polluting the namespace and also avoid all discussions on which
header file is best to include the definition of the PAGE_SIZE constant,
we instead pass it as a command-line definition.

We also hard-code its value because we want to ensure it is an unsigned,
rather than integer that the commented out code to get it from the
environment or a simple 4096 would default to.