Rename a function and declare it static
As it is used in the `swdsm.cpp` file only, there is no reason to expose
this function to the rest of the world.  Also, there is no reason not to
give it a name that better describes its current functionality, which is
crucial for correctness in the places where it is used.