Consistently add whitespaces in // comments (#127)
* Consistently add whitespaces in // comments

The rule that `cpplint` enforces is that // comments should be
preceeded by at least two spaces and followed by at least one.
Modify the code to follow it.

While at it, took the liberty to cleanup and/or remove some comments.

* Rename a function and declare it static

As it is used in the `swdsm.cpp` file only, there is no reason to expose
this function to the rest of the world.  Also, there is no reason not to
give it a name that better describes its current functionality, which is
crucial for correctness in the places where it is used.