Eliminate uses of the 4096 constant (#125)
This constant was used for the size of a page, but was hard-coded in
various places using various forms (4096, 4096UL, ...).  Also, various
header files used different definitions / names for it (`pagesize`,
`page_size`, `PAGESIZE`).

To avoid polluting the namespace and also avoid all discussions on which
header file is best to include the definition of the `PAGE_SIZE` constant,
we instead pass it as a command-line definition.

We also hard-code its value because we want to ensure it is an unsigned,
rather than integer that the commented out code to get it from the
environment or a simple 4096 would default to.

Addresses the biggest part of #49.  (Does not deal with CACHELINE.)