Fix external projects missing PAGE_SIZE
This commit introduces an automatically generated configuration
file config.hpp that contains the definitions of PAGE_SIZE and
CACHELINE based on values set in CMake configuration. This file
can be included through #include "config.hpp" from any file that
needs to access these macros. The set of possible CMake values
to choose between is restricted to only allow for selection of
values supported by the codebase.