Ensure that the reserved VM area fits
Due to changes in the Linux Virtual Memory layout, the reserved
virtual memory under MEMFD intruded on the area designated for
PIE loads and the heap on modern systems. This commit moves the
reserved virtual memory area to a safe region. For simplicity of
future debugging, the virtual memory regions of SHM and MEMFD are
also moved to the same start address.