Move VM reservation to fix MEMFD error (#142)
Due to changes in the Linux Virtual Memory layout, the reserved
virtual memory under MEMFD intruded on the area designated for
PIE loads and the heap on modern systems. This commit moves the
reserved virtual memory area to a safe region. For simplicity of
future debugging, the virtual memory regions of SHM and MEMFD
now use the same start address and size, and are defined in a 
separate header instead of replicated in each source file.

Since MEMFD should again be usable, it is added back to the GitHub
actions matrix. In order to run MEMFD we also enable the
always overcommit virtual memory accounting mode in the workflow.

This fixes #124