Fix 2 uninitialized variables (#147)
* Code rewrite of swdsm_argo_barrier() to avoid unitialized variable

New(er) compiler versions report that variable `barrierlockholder` may
be used unitialized in this function.  This was known for a while now,
but our (local) fix was unsatisfactory and was therefore not pushed to
the code base.  Motivated by #146, which is also not perfect as a fix,
we have decided to rewrite that code and eliminate the uses of this
variable, paying some cost in (small) code duplication.

Closes #146

Co-authored by: @davidklaftenegger

* Initialize variable to suppress warning